Finish CMS

+++ libs/core/src/main/kotlin/io/stereov/singularity/content/tag/service/TagService.kt
@@ -12,26 +12,28 @@ import io.stereov.singularity.content.tag.mapper.TagMapper
 import io.stereov.singularity.content.tag.model.TagDocument
 import io.stereov.singularity.content.tag.model.TagTranslation
 import io.stereov.singularity.content.tag.repository.TagRepository
+import io.stereov.singularity.database.core.service.CrudService
 import io.stereov.singularity.global.exception.model.DocumentNotFoundException
 import io.stereov.singularity.global.properties.AppProperties
-import io.stereov.singularity.global.util.getFieldContainsCriteria
+import io.stereov.singularity.global.util.CriteriaBuilder
 import jakarta.annotation.PostConstruct
-import kotlinx.coroutines.reactive.awaitFirstOrNull
 import kotlinx.coroutines.runBlocking
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.data.mongodb.core.ReactiveMongoTemplate
-import org.springframework.data.mongodb.core.find
-import org.springframework.data.mongodb.core.query.Query
 import org.springframework.stereotype.Service
 import java.util.*
 
 @Service
 class TagService(
-    private val repository: TagRepository,
-    private val reactiveMongoTemplate: ReactiveMongoTemplate,
+    override val repository: TagRepository,
+    override val reactiveMongoTemplate: ReactiveMongoTemplate,
     private val contentProperties: ContentProperties,
     private val tagMapper: TagMapper,
     private val appProperties: AppProperties,
-) {
+) : CrudService<TagDocument> {
+
+    override val collectionClazz = TagDocument::class.java
 
     @PostConstruct
     fun initializeTags() {
@@ -47,12 +49,11 @@ class TagService(
         }
     }
 
-    private val logger: KLogger
+    override val logger: KLogger
         get() = KotlinLogging.logger {}
 
     suspend fun create(req: CreateTagRequest): TagDocument {
         logger.debug { "Creating tag with key ${req.key}" }
-
         if (existsByKey(req.key)) throw TagKeyExistsException(req.key)
 
         return save(tagMapper.createTag(req))
@@ -64,12 +65,6 @@ class TagService(
         if (existsByKey(req.key)) throw TagKeyExistsException(req.key)
 
         return save(tagMapper.createTag(req))
-    }
-
-    suspend fun save(tag: TagDocument): TagDocument {
-        logger.debug { "Saving tag with key \"${tag.key}\"" }
-
-        return repository.save(tag)
     }
 
     suspend fun findByKey(key: String): TagDocument {
@@ -85,19 +80,23 @@ class TagService(
         return repository.existsByKey(name)
     }
 
-    suspend fun findByNameContains(substring: String, locale: Locale?): List<TagDocument> {
-        logger.debug { "Finding tags with name containing \"$substring\"" }
+    suspend fun findAllPaginated(
+        pageable: Pageable,
+        key: String?,
+        name: String?,
+        description: String?,
+        locale: Locale?
+    ): Page<TagDocument> {
+        logger.debug { "Finding tags" }
 
         val actualLocale = locale ?: appProperties.locale
+        val criteria = CriteriaBuilder()
+            .fieldContains(TagDocument::key, key)
+            .fieldContains(TagTranslation::name, name, actualLocale)
+            .fieldContains(TagTranslation::description, description, actualLocale)
+            .build()
 
-        val field = "${TagDocument::translations.name}.$actualLocale.${TagTranslation::name.name}"
-
-        val criteria = getFieldContainsCriteria(field, substring)
-
-        return reactiveMongoTemplate.find<TagDocument>(Query.query(criteria))
-            .collectList()
-            .awaitFirstOrNull()
-            ?: emptyList()
+        return findAllPaginated(pageable, criteria)
     }
 
     suspend fun updateTag(key: String, req: UpdateTagRequest): TagDocument {
@@ -105,22 +104,21 @@ class TagService(
 
         val tag = findByKey(key)
         val updatedTranslations = mutableMapOf<Locale, TagTranslation>()
+        val actualLocale = req.locale ?: appProperties.locale
 
-        req.translations.forEach { (locale, updateReq) ->
-            val existing = tag.translations[locale]
+        val existing = tag.translations[actualLocale]
 
-            if (existing != null) {
-                updatedTranslations[locale] = TagTranslation(
-                    updateReq.name ?: existing.name,
-                    updateReq.description ?: existing.description
-                )
-            } else {
-                updatedTranslations[locale] = TagTranslation(
-                    updateReq.name
-                        ?: throw InvalidUpdateTagRequest("Failed to add new translation $locale for tag \"$key\": tag name not specified"),
-                    updateReq.description ?: ""
-                )
-            }
+        if (existing != null) {
+            updatedTranslations[actualLocale] = TagTranslation(
+                req.name ?: existing.name,
+                req.description ?: existing.description
+            )
+        } else {
+            updatedTranslations[actualLocale] = TagTranslation(
+                req.name
+                    ?: throw InvalidUpdateTagRequest("Failed to add new translation $actualLocale for tag \"$key\": tag name not specified"),
+                req.description ?: ""
+            )
         }
 
         tag.translations.putAll(updatedTranslations)
@@ -128,15 +126,10 @@ class TagService(
         return save(tag)
     }
 
-    suspend fun deleteByKey(key: String): Boolean {
+    suspend fun deleteByKey(key: String) {
         logger.debug { "Deleting tag with key \"$key\"" }
 
-        return repository.deleteByKey(key)
-    }
-
-    suspend fun deleteAll() {
-        logger.debug { "Deleting all tags" }
-
-        repository.deleteAll()
+        if (!existsByKey(key)) throw DocumentNotFoundException("No tag with key $key found")
+        repository.deleteByKey(key)
     }
 }
